Refactor :)
