+++ lua/open_browser_git.lua
@@ -25,54 +25,8 @@ local M = {
   },
 }
 
--- Remove duplicate items from a list-like table. Does not modify `t`, may
--- shuffle items, discards keys.
-local function tbl_uniq(t, fn)
-  local t2 = {}
-  for _, value in ipairs(t) do
-    t2[fn(value)] = value
-  end
-  return vim.tbl_values(t2)
-end
-
-local function parse_git_remote_url(url)
-  -- User, repo, whitespace.
-  -- NB: We trim a trailing `.git` from the repo.
-  local user_repo_pattern = "([^/]+)/([^/]+)%s"
-  -- NB: We trim a leading `git@` or other username from the hostname.
-  local patterns = {
-    -- NB: This first pattern also matches ssh://git@... URLs.
-    "git@([^:/]+)[:/]" .. user_repo_pattern, -- ssh
-    "%s([^:/]+)[:/]" .. user_repo_pattern, -- ssh
-    "ssh://([^/]+)/" .. user_repo_pattern, -- ssh
-    "git://([^/]+)/" .. user_repo_pattern, -- git
-    "https?://([^/]+)/" .. user_repo_pattern, -- http(s)
-  }
-  for _, pattern in ipairs(patterns) do
-    local host, user, repo = url:match(pattern)
-    if host ~= nil then
-      local remote_name = url:match("^([^\t]+)\t")
-      return require("open_browser_git.repo"):new {
-        host = host:gsub("^([^@]+)@", ""),
-        user = user,
-        repo = repo:gsub("%.git$", ""),
-        remote_name = remote_name,
-        flavor_patterns = M._config.flavor_patterns,
-      }
-    end
-  end
-end
-
-function M.parse_git_remote(lines, callback)
-  local repos = {}
-  for _, line in ipairs(lines) do
-    -- Can I simplify this to skip the nil check?
-    local repo = parse_git_remote_url(line)
-    if repo ~= nil then
-      table.insert(repos, repo)
-    end
-  end
-  repos = tbl_uniq(repos, require("open_browser_git.repo").display)
+function M.pick_remote(path, callback)
+  local repos = path:list_remotes()
   table.sort(repos, function(a, b)
     if a.remote_name == "origin" and b.remote_name ~= "origin" then
       -- Sort 'origin' remotes first.
@@ -92,10 +46,6 @@ function M.parse_git_remote(lines, callback)
   end
 end
 
-function M.get_git_repo(path, callback)
-  M.parse_git_remote(path:git({ "remote", "-v" }).stdout, callback)
-end
-
 -- If the `path` is given, open that file in a browser.
 --
 -- If the `path` is absent, open the repo's homepage.
@@ -103,7 +53,7 @@ end
 -- open_git(path: string|nil, options: {lines: {line1: int, line2: int}|nil}|nil)
 function M.open_git(path, options)
   path = require("open_browser_git.path"):new(path)
-  M.get_git_repo(path, function(repo)
+  M.pick_remote(path, function(repo)
     -- TODO: Find the most recent _pushed_ commit.
     local commit = path:git({ "rev-parse", "HEAD" }).stdout[1]
     local url = repo:url_for_file(path:relative_to_root(), commit, options)
